fix: duplicate frontmatter thanks to new Obsidian API, I can readd editing with regex on the frontmatter without duplicating it
